+++ errors_test.go
@@ -5,7 +5,7 @@ import "fmt"
 func ExampleMalformed_Error() {
 	var e Malformed
 	e.SetPacket(NewConnect())
-	e.SetReason("missing data")
+	e.SetReasonString("missing data")
 	fmt.Println(e.Error())
 	// output:
 	// malformed *mq.Connect: missing data